Allow renting film from existing inventory

+++ src/main/java/com/olga_o/dto/RentalReturnDto.java
@@ -1,6 +1,6 @@
 package com.olga_o.dto;
 
 public class RentalReturnDto {
-    public int rentalId;
-    public int customerId;
+    public short rentalId;
+    public short customerId;
 }
